% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{apply_wght_conv_fct}
\alias{apply_wght_conv_fct}
\title{Apply weight conversion factors to HCES data}
\usage{
apply_wght_conv_fct(
  hces_df,
  conv_fct_df,
  factor_col = "factor",
  measure_id_col = "measure_id",
  wt_kg_col = "wt_kg",
  cons_qnty_col = "cons_quant",
  allowDuplicates = FALSE
)
}
\arguments{
\item{hces_df}{A dataframe containing the HCES data.}

\item{conv_fct_df}{A dataframe containing the conversion factors.}

\item{factor_col}{The name of the column in `conv_fct_df` containing the conversion factors.}

\item{measure_id_col}{The name of the column in both dataframes containing the measure IDs.}

\item{wt_kg_col}{The name of the column to store the weight in kg.}

\item{cons_qnty_col}{The name of the column in `hces_df` containing the consumption quantity.}

\item{allowDuplicates}{A logical value indicating whether to allow duplicates in the measure_id column of the 
conversion factors dataframe. Default is FALSE.}
}
\value{
A dataframe containing the original HCES data with additional columns for the conversion factor and weight 
in kg.
}
\description{
This function applies weight conversion factors to the HCES data based on the conversion factors dataframe. 
It merges the HCES data and conversion factors dataframe on the `measure_id` column, and calculates the weight in kg 
for each food item using the conversion factor and consumption quantity. The resulting dataframe contains the 
original HCES data with additional columns for the conversion factor and weight in kg.
}
\note{
The function checks if the required columns are present in the dataframes and throws an error if they are not. 
It also checks if there are duplicates in the measure_id column of the conversion factors dataframe and throws an 
error if there are any.
}
\examples{
\dontrun{
apply_wght_conv_fct(hces_df, conv_fct_df, factor_col = "factor", measure_id_col = "measure_id", 
wt_kg_col = "wt_kg", cons_qnty_col = "cons_quant", allowDuplicates = FALSE)
}
}
