% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{assign_unit_codes}
\alias{assign_unit_codes}
\title{Assign unit codes to a data frame based on a unit codes CSV file}
\usage{
assign_unit_codes(
  data,
  country,
  survey,
  unit_code_col,
  unit_name_col,
  unit_codes_csv = NULL
)
}
\arguments{
\item{data}{A data frame to modify.}

\item{country}{A character string specifying the country of the data.}

\item{survey}{A character string specifying the survey of the data.}

\item{unit_code_col}{A character string specifying the name of the column in the data frame to store the unit codes.}

\item{unit_name_col}{A character string specifying the name of the column in the data frame to match the unit names.}

\item{unit_codes_csv}{A character string specifying the file path of the unit codes CSV file.
If NULL, the internal data hcesNutR::unit_codes will be used.}
}
\value{
A modified data frame with unit codes assigned.
}
\description{
This function assigns unit codes to a data frame based on a unit codes CSV file. 
The CSV file must contain a column with unit names and a column with corresponding unit codes.
The function removes special characters from the unit names before matching them to the data frame.
}
\examples{
data <- data.frame(cons_unit_name = c("kg", "g", "lb", "oz"))
assign_unit_codes(data, "USA", "NHANES", NULL, "unit_code", "cons_unit_name")
}
