% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{check_conv_fct}
\alias{check_conv_fct}
\title{Check for missing conversion factors}
\usage{
check_conv_fct(hces_df, conv_fct_df)
}
\arguments{
\item{hces_df}{A dataframe containing the `measure_id` column.}

\item{conv_fct_df}{A dataframe containing the `measure_id` column.}
}
\value{
A dataframe `missing_conv_fct_df` containing the missing `measure_id` values.
}
\description{
This function checks if the `measure_id` column is present in both input dataframes `hces_df` and `conv_fct_df`. If `measure_id` is not present in either of the dataframes, the function throws an error. The function then filters out rows in `hces_df` that are not in `conv_fct_df` based on the `measure_id` column. It returns a dataframe `missing_conv_fct_df` that contains the missing `measure_id` values.
}
\note{
Use `View()` to see the missing conversion factors.
}
\examples{
\dontrun{
check_conv_fct(hces_df, conv_fct_df)
}
}
