% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{check_hces_names}
\alias{check_hces_names}
\title{Compare column names in HCES data}
\usage{
check_hces_names(data, original_data_df)
}
\arguments{
\item{data}{dataframe . The data file as a dataframe.}

\item{original_data_df}{dataframe . The original data file as a dataframe.}
}
\value{
dataframe: The data file with renamed columns, converted data types and a table of the changes to the console.
}
\description{
This function compares the column names in the hces data file to the standard MAPS column names. It print the changes to the console.
}
