% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined_food_list}
\alias{combined_food_list}
\title{Food Names and Codes Supported by this Package}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 292 rows and 6 columns.
}
\usage{
combined_food_list
}
\description{
A data frame containing non-standard food names and codes of surveys supported by this package.
}
\keyword{datasets}
