% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{create_dta_labels}
\alias{create_dta_labels}
\title{Create labels for Stata data}
\usage{
create_dta_labels(data)
}
\arguments{
\item{data}{dataframe . The data file as a dataframe.}
}
\value{
dataframe: The data file with labels for all <dbl+lbl> columns.
}
\description{
This function creates labels for all <dbl+lbl> columns in data imported from Stata data.
}
\note{
The function uses the haven package to create the labels and values columns.

The function will only create the labels and values columns if they do not already exist. It will not overwrite existing columns.
}
