% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{create_matches_csv}
\alias{create_matches_csv}
\title{Create Matches CSV}
\usage{
create_matches_csv()
}
\value{
A Shiny application.
}
\description{
This function creates a Shiny application that allows users to match non-standard food items to standard ones.
The matches are based on the closest string match. The user can manually adjust the matches in the UI.
The matched data can be added to the R environment or downloaded as a CSV file.
}
\examples{
\dontrun{
create_matches_csv()
}
}
