% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{food_list_df}
\alias{food_list_df}
\title{Food Item Names and Codes Supported by this Package}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 351 rows and 8 columns.
}
\usage{
food_list_df
}
\description{
A data frame containing both standard and non-standard food item names and codes of surveys supported by this package.
}
\keyword{datasets}
