% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{get_hces_mappings}
\alias{get_hces_mappings}
\title{Get column mapping for a given country and survey}
\usage{
get_hces_mappings(country_name, survey_name)
}
\arguments{
\item{country_name}{character The country name in ISO3 format.}

\item{survey_name}{character The survey name in short form.}
}
\value{
dataframe: The column mapping for the given country and survey.
}
\description{
This function returns the column mapping for a given country and survey
}
