% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hh_mod_a_filt_vMAPS}
\alias{hh_mod_a_filt_vMAPS}
\title{Module A: Household Identification (household level data)}
\format{
A data frame
}
\usage{
hh_mod_a_filt_vMAPS
}
\description{
Randomly Generated Samaple Data Resembling Data collected through Household Questionnaire, Module A: Household Identification (household level data)
- This module household identifiers, the sample weights, information on household location, date of interview, supervisor and enumerator codes. Additionally, this module contains filters for subsequent modules.
}
\keyword{datasets}
