% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{match_food_names}
\alias{match_food_names}
\alias{match_food_names_v2}
\title{Match food names to a standard list of food items}
\usage{
match_food_names(
  data,
  country,
  survey,
  food_name_col,
  food_code_col,
  matches_csv = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{data}{A data frame containing the food names and codes to be matched.}

\item{country}{A character string specifying the country for which the food items should be matched.}

\item{survey}{A character string specifying the survey for which the food items should be matched.}

\item{food_name_col}{A character string specifying the name of the column in \code{data} containing the food names.}

\item{food_code_col}{A character string specifying the name of the column in \code{data} containing the food codes.}

\item{matches_csv}{A character string specifying the path to a user-provided CSV file containing the food list to be used for matching. If \code{NULL}, the function uses an internal food list.}

\item{overwrite}{A logical value indicating whether to overwrite the original food name and code columns in \code{data} with the matched food names and codes. If \code{TRUE}, the original columns are overwritten. If \code{FALSE}, new columns are created for the matched food names and codes.}
}
\value{
A data frame with the matched food names and codes.
}
\description{
This function matches food names to a standard list of food items and returns a data frame with the matched food names and codes. The function can use an internal food list or a user-provided food list to perform the matching. The function also provides diagnostics on the number of rows with no matches found and the list of foods with no matches found.
}
\examples{
\dontrun{
# Load the MWI data
data("mwi_data", package = "hcesNutR")

# Match the food names to the standard list of food items
matched_data <- match_food_names_v2(data = mwi_data, 
                                    country = "MWI", 
                                    survey = "IHS5", 
                                    food_name_col = "food_item_name", 
                                    food_code_col = "food_item_code")
}

}
