% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{match_food_names_v1}
\alias{match_food_names_v1}
\title{Match food names in a dataset to an internal standard food list}
\usage{
match_food_names_v1(
  data,
  country,
  survey,
  food_name_col,
  match_confidence = 90
)
}
\arguments{
\item{data}{A data frame containing the food names to be matched.}

\item{country}{A character string specifying the country for which the food list should be generated.}

\item{survey}{A character string specifying the survey for which the food list should be generated.}

\item{food_name_col}{A character string specifying the name of the column in \code{data} that contains the food names to be matched.}

\item{match_confidence}{A numeric value specifying the minimum match confidence level required for a match to be considered valid. Defaults to 90.}
}
\value{
A data frame with the following columns:
\item{food_item_code}{A character string containing the standard food item code for the matched food item.}
\item{standard_original_food_name}{A character string containing the standard food name for the matched food item.}
\item{food_match_source}{A character string containing the source of the matched food item.}
\item{food_match_confidence}{A numeric value containing the match confidence level for the matched food item.}
}
\description{
This function matches food names in a dataset to a standard food list based on the closest string distance match. It then appends the food item code, standard food name, source and match confidence level to the dataset.
}
