% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{non_standard_food_list}
\alias{non_standard_food_list}
\title{Non Standard Food Names and Codes}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 164 rows and 5 columns.
}
\usage{
non_standard_food_list
}
\description{
A data frame containing non-standard food names and codes of surveys supported by this package.
}
\keyword{datasets}
