% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{remove_spec_char_v1}
\alias{remove_spec_char_v1}
\title{Remove special characters from a column in a dataframe}
\usage{
remove_spec_char_v1(data, column, keep_parenthesis = TRUE)
}
\arguments{
\item{data}{dataframe. The dataframe to modify.}

\item{column}{character. The name of the column to modify.}

\item{keep_parenthesis}{logical. If TRUE, parenthesis are kept in the column. If FALSE, parenthesis are removed. Default is TRUE.}
}
\value{
dataframe. The modified dataframe.
}
\description{
This function removes special characters from a specified column in a dataframe. It also adds a period after 'NO' if it is a whole word and not already followed by a period.
}
\examples{
data <- data.frame(x = c("NO", "YES (maybe)", "NO (definitely)"))
remove_spec_char_v1(data, "x", FALSE)
}
