% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{remove_unconsumed}
\alias{remove_unconsumed}
\title{Remove unconsumed food item records}
\usage{
remove_unconsumed(data, consCol = "consYN", consVal = "YES")
}
\arguments{
\item{data}{dataframe . The data file as a dataframe.}

\item{consCol}{character . The name of the column that indicates whether the food item was consumed or not.}
}
\description{
This function removes the unconsumed food item records from the data file.
}
\note{
the read_in_data() function should be used first to import and format the required data.
}
