% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{rename_hces}
\alias{rename_hces}
\title{Rename data columns in HCES data}
\usage{
rename_hces(data, country_name, survey_name)
}
\arguments{
\item{data}{dataframe . The data file as a dataframe.}

\item{column_mapping}{dataframe . The column mapping file as a dataframe.}
}
\value{
dataframe: The data file with renamed columns.
}
\description{
This function renames the columns in the hces data file to the standard MAPS column names.
}
\note{
the read_in_data() function should be used first to import and format the required data.
}
