% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{replace_values}
\alias{replace_values}
\title{Replace values in a column of a dataframe}
\usage{
replace_values(data, targetColumn, secondaryColumn, string2search)
}
\arguments{
\item{data}{The dataframe to modify.}

\item{targetColumn}{The name of the column to modify.}

\item{secondaryColumn}{The name of the column to extract values from and replace in targetColumn.}

\item{string2search}{The value to search. It can be the whole string or a part of it.}
}
\value{
The modified dataframe.
}
\description{
This function replaces values in a column of a dataframe with a new value.
}
\examples{
data <- data.frame(x = c("a", "b", "c"), y = c("d", "e", "f"))
replace_values(data, "x", "y", "b")
}
