% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{rm_special_chars_v2}
\alias{rm_special_chars_v2}
\title{Remove special characters and spaces from a string vector and capitalize the result}
\usage{
rm_special_chars_v2(str_vec)
}
\arguments{
\item{str_vec}{A vector of strings to be processed}
}
\value{
A vector of strings where each string is collapsed into one string and capitalized
}
\description{
This function takes a vector of strings as input, removes all special characters and spaces from each string, and returns a new vector of strings where each string is collapsed into one string and capitalized.
}
\examples{
rm_special_chars_v2(c("PACKET (LARGE)", "BAG (SMALL)", "BOX (MEDIUM)"))

}
