% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{split_dta}
\alias{split_dta}
\title{Split a labelled variable into two new variables}
\usage{
split_dta(data, split_var, val_to, lab_to, drop_split_var = TRUE)
}
\arguments{
\item{data}{A data frame containing the input variable.}

\item{split_var}{A character string specifying the name of labelled the input variable to split.}

\item{val_to}{A character string specifying the name of the new variable to contain the numeric values.}

\item{lab_to}{A character string specifying the name of the new variable to contain the character labels.}

\item{drop_split_var}{A logical value indicating whether to drop the original variable (default is TRUE).}
}
\value{
A data frame with the new variables added and the original variable dropped (if specified).
}
\description{
This function splits a labelled variable into two new variables: one containing
the numeric values and one containing the character labels. The new variables
are added to the input data frame and the original variable is dropped by default.
}
\examples{
library(haven)
df <- data.frame(
  HHID = c(rep("hh01",5),rep("hh02",5)),
  food_item = (rep(101:105,2)),
  consYN = sample(1:2,10,replace = TRUE)
) 

# Add value labels
df$food_item <- labelled(df$food_item, c("maize" = 101, "wheat" = 102, 
"rice" = 103, "meat" = 104, "fish" = 105))
df$consYN <- labelled(df$consYN, c("Yes" = 1, "No" = 2))

# Print data frame
df

# Split the food_item column into two new columns
split_dta(df, "food_item", "food_item_id", "food_item_lab")

}
