% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_food_list}
\alias{standard_food_list}
\title{Standard Food Names and Codes}
\format{
An object of class \code{data.frame} with 142 rows and 4 columns.
}
\usage{
standard_food_list
}
\description{
A data frame containing food names and codes of surveys supported by this package.
}
\keyword{datasets}
