% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard_name_mappings_pairs}
\alias{standard_name_mappings_pairs}
\title{Standard HCES Variable Name Mappings}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 32 rows and 5 columns.
}
\usage{
standard_name_mappings_pairs
}
\description{
A data frame containing standard variable name mapping pairs for the IHS5 survey.
}
\keyword{datasets}
