% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unit_names_n_codes_df}
\alias{unit_names_n_codes_df}
\title{Food Unit Names and Unit Codes Supported by this Package}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 208 rows and 6 columns.

A data frame with 3 columns and 4 rows
\describe{
\item{country}{Country name}
\item{survey}{Survey name}
\item{unit_name}{Name of the food unit}
\item{unit_code}{Code of the food unit}
\entries_count{Number of survey responses with this unit name}
\priority{Priority of the unit name (1 being the highest)}}
}
\usage{
unit_names_n_codes_df

unit_names_n_codes_df
}
\description{
A data frame containing both standard and non-standard food unit names and unit codes of surveys supported by this package.

A data frame containing both standard and non-standard food unit names and unit codes of surveys supported by this package.
}
\examples{
unit_names_n_codes_df
}
\keyword{datasets}
