% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_hces.R
\name{which_labelled}
\alias{which_labelled}
\title{Identify labelled variables in a data frame}
\usage{
which_labelled(data)
}
\arguments{
\item{data}{A data frame to check for labelled variables.}
}
\value{
A character vector of the names of the labelled variables in the input data frame.
}
\description{
This function takes a data frame as input and returns a character vector of the names of the variables in the data frame that are labelled using the `haven` package.
}
\examples{
library(haven)
df <- data.frame(
  HHID = c(rep("hh01",5),rep("hh02",5)),
  food_item = (rep(101:105,2)),
  consYN = sample(1:2,10,replace = TRUE)
) 

# Add value labels
df$food_item <- labelled(df$food_item, c("maize" = 101, "wheat" = 102, "rice" = 103, "meat" = 104, "fish" = 105))
df$consYN <- labelled(df$consYN, c("Yes" = 1, "No" = 2))

# Print data frame
df

# Check for labelled variables
which_labelled(df)


}
